requirements.txt
asgiref==3.11.0
bandit==1.9.2
certifi==2025.11.12
cfgv==3.5.0
charset-normalizer==3.4.4
distlib==0.4.0
Django==5.2.9
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
filelock==3.20.1
flake8==7.3.0
gunicorn==23.0.0
identify==2.6.15
idna==3.11
Markdown==3.10
markdown-it-py==4.0.0
mccabe==0.7.0
mdurl==0.1.2
nodeenv==1.10.0
packaging==25.0
pillow==12.0.0
platformdirs==4.5.1
pre_commit==4.5.1
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
rich==14.2.0
sqlparse==0.5.5
stevedore==5.6.0
typing_extensions==4.15.0
urllib3==2.6.2
virtualenv==20.35.4
